Restructure Packages by function rather than model class
Due to naming conflict between Projections with new Package structure
InlineDinghyClass renamed to DinghyInlineDinghyClass and
RaceInlineDinghyClass added
