+++ src/main/java/com/bginfosys/dinghyracing/model/DinghyClass.java
@@ -1,0 +1,46 @@
+package com.bginfosys.dinghyracing.model;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.Version;
+
+import javax.validation.constraints.NotNull;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+
+@Entity
+public class DinghyClass {
+	
+	private @Id @GeneratedValue Long id;
+	private @Version @JsonIgnore Long version;
+	
+	@NotNull
+	@Column(unique=true)
+	private String name;
+
+	//Required by JPA
+	//Not recommended by Spring Data
+	public DinghyClass() {}
+	
+	public DinghyClass(String name) {
+		this.name = name;	
+	}
+	
+	public Long getId() {
+		return id;
+	}
+	
+	public void setId(Long id) {
+		this.id = id;
+	}
+	
+	public String getName() {
+		return name;
+	}
+
+	public void setName(String name) {
+		this.name = name;
+	}
+}
