+++ src/main/java/com/bginfosys/dinghyracing/model/DinghyInlineDinghyClass.java
@@ -1,0 +1,11 @@
+package com.bginfosys.dinghyracing.model;
+
+import org.springframework.data.rest.core.config.Projection;
+
+@Projection(name = "dinghyInlineDinghyClass", types = {Dinghy.class})
+public interface DinghyInlineDinghyClass {
+	
+	String getSailNumber();
+	
+	DinghyClass getDinghyClass();
+}
