+++ src/main/java/com/bginfosys/dinghyracing/model/DinghyNotFoundException.java
@@ -1,0 +1,11 @@
+package com.bginfosys.dinghyracing.model;
+
+public class DinghyNotFoundException extends RuntimeException {
+	
+	private static final long serialVersionUID = 1L;
+	
+	public DinghyNotFoundException(Long id) {
+		super("Could not find dinghy with id " + id);
+	}
+
+}
