+++ src/main/java/com/bginfosys/dinghyracing/model/RaceNotFoundException.java
@@ -1,0 +1,10 @@
+package com.bginfosys.dinghyracing.model;
+
+public class RaceNotFoundException extends RuntimeException {
+	
+	private static final long serialVersionUID = 1L;
+
+	public RaceNotFoundException(Long id) {
+		super("Could not find race with id " + id);
+	}
+}
