+++ src/main/java/com/bginfosys/dinghyracing/persistence/DinghyClassRepository.java
@@ -1,0 +1,21 @@
+package com.bginfosys.dinghyracing.persistence;
+
+import org.springframework.data.repository.PagingAndSortingRepository;
+import org.springframework.data.repository.query.Param;
+import org.springframework.data.rest.core.annotation.RepositoryRestResource;
+
+import com.bginfosys.dinghyracing.model.DinghyClass;
+
+@RepositoryRestResource(path = "dinghyclasses") //, exported = false)
+public interface DinghyClassRepository extends PagingAndSortingRepository<DinghyClass, Long> {
+
+	@Override
+	DinghyClass save(@Param("dinghyClass") DinghyClass dinghyClass);
+	
+	@Override
+	void deleteById(@Param("id") Long id);
+
+	@Override
+	void delete(@Param("dinghyClass") DinghyClass dinghyClass);
+	
+}
