--- src/main/java/com/bginfosys/dinghyracing/race/RaceNotFoundAdvice.java
@@ -1,18 +1,0 @@
-package com.bginfosys.dinghyracing.race;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class RaceNotFoundAdvice {
-	
-	@ResponseBody
-	@ExceptionHandler(RaceNotFoundException.class)
-	@ResponseStatus(HttpStatus.NOT_FOUND)
-	String raceNotFoundHandler(RaceNotFoundException ex) {
-		return ex.getMessage();
-	}
-}
