--- src/main/java/com/bginfosys/dinghyracing/race/RaceNotFoundException.java
@@ -1,10 +1,0 @@
-package com.bginfosys.dinghyracing.race;
-
-public class RaceNotFoundException extends RuntimeException {
-	
-	private static final long serialVersionUID = 1L;
-
-	public RaceNotFoundException(Long id) {
-		super("Could not find race with id " + id);
-	}
-}
