--- src/main/java/com/bginfosys/dinghyracing/race/RaceWebSocketConfiguration.java
@@ -1,25 +1,0 @@
-package com.bginfosys.dinghyracing.race;
-
-import org.springframework.messaging.simp.config.MessageBrokerRegistry;
-import org.springframework.stereotype.Component;
-import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
-import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
-import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerConfigurer;
-
-@Component
-@EnableWebSocketMessageBroker
-public class RaceWebSocketConfiguration implements WebSocketMessageBrokerConfigurer {
-
-		static final String MESSAGE_PREFIX = "/topic";
-		
-		@Override
-		public void registerStompEndpoints(StompEndpointRegistry registry) {
-			registry.addEndpoint("/dinghyracing/messages").setAllowedOriginPatterns("*").withSockJS();
-		}
-		
-		@Override
-		public void configureMessageBroker(MessageBrokerRegistry registry) {
-			registry.enableSimpleBroker(MESSAGE_PREFIX);
-			registry.setApplicationDestinationPrefixes("/app");
-		}
-}
