+++ src/test/java/com/bginfosys/dinghyracing/model/RaceNotFoundExceptionTests.java
@@ -1,0 +1,5 @@
+package com.bginfosys.dinghyracing.model;
+
+public class RaceNotFoundExceptionTests {
+
+}
