+++ src/test/java/com/bginfosys/dinghyracing/persistence/RaceRepositoryTests.java
@@ -1,0 +1,5 @@
+package com.bginfosys.dinghyracing.persistence;
+
+public class RaceRepositoryTests {
+
+}
