--- src/test/java/com/bginfosys/dinghyracing/race/InlineDinghyClassTests.java
@@ -1,13 +1,0 @@
-package com.bginfosys.dinghyracing.race;
-
-import org.junit.jupiter.api.Test;
-
-import java.time.LocalDateTime;
-
-import com.bginfosys.dinghyracing.dinghyclass.DinghyClass;
-
-public class InlineDinghyClassTests {
-
-	//Create an instance of the interface and confirm it returns a dinghy class
-	//Does this require other elements? (It is a projection over Race used by another component in Spring Data Rest
-}
