--- src/test/java/com/bginfosys/dinghyracing/race/RaceRepositoryTests.java
@@ -1,5 +1,0 @@
-package com.bginfosys.dinghyracing.race;
-
-public class RaceRepositoryTests {
-
-}
