+++ src/test/java/com/bginfosys/dinghyracing/web/rest/InlineDinghyClassTests.java
@@ -1,0 +1,13 @@
+package com.bginfosys.dinghyracing.web.rest;
+
+import org.junit.jupiter.api.Test;
+
+import com.bginfosys.dinghyracing.model.DinghyClass;
+
+import java.time.LocalDateTime;
+
+public class InlineDinghyClassTests {
+
+	//Create an instance of the interface and confirm it returns a dinghy class
+	//Does this require other elements? (It is a projection over Race used by another component in Spring Data Rest
+}
